+++ src/components/Wizards/CreateManagedControlPlane/CreateManagedControlPlaneV2WizardContainer.cy.tsx
@@ -1,8 +1,8 @@
 import '@ui5/webcomponents-cypress-commands';
 import { useAuthOnboarding } from '../../../spaces/onboarding/auth/AuthContextOnboarding.tsx';
 import { ManagedControlPlaneV2 } from '../../../spaces/onboarding/types/ControlPlane.ts';
-import { useCreateManagedControlPlaneV2GraphQL } from '../../../spaces/controlPlaneV2/hooks/useCreateManagedControlPlaneV2GraphQL.ts';
-import { useUpdateManagedControlPlaneV2GraphQL } from '../../../spaces/controlPlaneV2/hooks/useUpdateManagedControlPlaneV2GraphQL.ts';
+import { useCreateControlPlaneV2GraphQL } from '../../../spaces/controlPlaneV2/hooks/useCreateControlPlaneV2GraphQL.ts';
+import { useUpdateControlPlaneV2GraphQL } from '../../../spaces/controlPlaneV2/hooks/useUpdateControlPlaneV2GraphQL.ts';
 import type { McpV2Input } from '../../../spaces/mcp/schemas/mcpV2Input.schema.ts';
 import { CreateManagedControlPlaneV2WizardContainer } from './CreateManagedControlPlaneV2WizardContainer.tsx';
 
@@ -24,7 +24,7 @@ describe('CreateManagedControlPlaneV2WizardContainer', () => {
     user: { email: 'user@example.com' },
   })) as typeof useAuthOnboarding;
 
-  const fakeUseCreateMcp: typeof useCreateManagedControlPlaneV2GraphQL = () => ({
+  const fakeUseCreateMcp: typeof useCreateControlPlaneV2GraphQL = () => ({
     createMcp: async (input: McpV2Input) => {
       createPayload = input;
       return mockMutationResult(input);
@@ -33,7 +33,7 @@ describe('CreateManagedControlPlaneV2WizardContainer', () => {
     error: undefined,
   });
 
-  const fakeUseUpdateMcp: typeof useUpdateManagedControlPlaneV2GraphQL = () => ({
+  const fakeUseUpdateMcp: typeof useUpdateControlPlaneV2GraphQL = () => ({
     updateMcp: async (input: McpV2Input) => {
       updatePayload = input;
       return mockMutationResult(input);
@@ -219,7 +219,7 @@ describe('CreateManagedControlPlaneV2WizardContainer', () => {
   });
 
   it('shows an error dialog when updateMcp throws in edit mode', () => {
-    const fakeFailingUpdate: typeof useUpdateManagedControlPlaneV2GraphQL = () => ({
+    const fakeFailingUpdate: typeof useUpdateControlPlaneV2GraphQL = () => ({
       updateMcp: async () => {
         throw new Error('Network error');
       },
