+++ src/spaces/controlPlaneV2/helpers/controlPlaneV2GraphQLInput.ts
