+++ src/spaces/controlPlaneV2/hooks/useDeleteControlPlaneV2GraphQL.ts
@@ -2,9 +2,9 @@ import { useCallback } from 'react';
 import { useMutation } from '@apollo/client/react';
 import { useToast } from '../../../context/ToastContext.tsx';
 import { useTranslation } from 'react-i18next';
-import { DeleteManagedControlPlaneV2Mutation } from './useDeleteManagedControlPlaneV2Mutation.ts';
+import { DeleteManagedControlPlaneV2Mutation } from './useDeleteControlPlaneV2Mutation.ts';
 
-export function useDeleteManagedControlPlaneV2GraphQL(namespace: string, name: string) {
+export function useDeleteControlPlaneV2GraphQL(namespace: string, name: string) {
   const { t } = useTranslation();
   const toast = useToast();
   const [deleteMutation] = useMutation(DeleteManagedControlPlaneV2Mutation);
