+++ src/spaces/controlPlaneV2/hooks/useUpdateControlPlaneV2GraphQL.ts
@@ -1,11 +1,11 @@
 import { useMutation } from '@apollo/client/react';
 import { useCallback } from 'react';
 import { z } from 'zod';
-import { buildMcpV2GraphQLInput } from '../helpers/mcpV2GraphQLInput.ts';
+import { buildMcpV2GraphQLInput } from '../helpers/controlPlaneV2GraphQLInput.ts';
 import { McpV2Input, McpV2InputSchema } from '../../mcp/schemas/mcpV2Input.schema.ts';
-import { UpdateManagedControlPlaneV2Mutation } from './useUpdateManagedControlPlaneV2Mutation.ts';
+import { UpdateManagedControlPlaneV2Mutation } from './useUpdateControlPlaneV2Mutation.ts';
 
-export function useUpdateManagedControlPlaneV2GraphQL() {
+export function useUpdateControlPlaneV2GraphQL() {
   const [updateMutation, { loading, error }] = useMutation(UpdateManagedControlPlaneV2Mutation);
 
   const updateMcp = useCallback(
